Создал счётик комментариев и их отрисовку по нажатию
