+++ js/gallery.js
@@ -1,5 +1,5 @@
 import { renderThumbmails } from './thumbnail.js';
-import { showBigPicture } from './big-picture.js';
+import { showBigPicture } from './picture.js';
 
 const container = document.querySelector('.pictures');
